ajustando swagger e security

+++ src/main/java/io/github/humbertoluiz/domain/repository/ClienteRepository.java
@@ -11,7 +11,7 @@ import io.github.humbertoluiz.domain.entity.Cliente;
 public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 	
 	@Query(" select c from Cliente c left join fetch c.pedidos where c.id = :id ")
-	Cliente findClienteFetchPedidos( @Param("id") Integer id );
+	Cliente findClienteFetchPedidos( @Param("id") Long id );
 
 	
 }
